+++ jovo-integrations/jovo-platform-googleassistant/test/FrameworkBase.test.ts
@@ -17,6 +17,7 @@ beforeEach(() => {
     t = ga.makeTestSuite();
 });
 
+
 describe('test request types', () => {
     test('test launch', async (done) => {
         app.setHandler({
@@ -699,6 +700,49 @@ describe('test handleOnRequest', () => {
         });
     });
 
+    test('ON_REQUEST return and skip intent handling', async (done) => {
+
+        app.setHandler({
+            ON_REQUEST() {
+                return this.tell('ON_REQUEST');
+            },
+            LAUNCH() {
+                // skip this
+                this.tell('LAUNCH');
+            },
+        });
+        const launchRequest:JovoRequest = await t.requestBuilder.launch();
+
+        app.handle(ExpressJS.dummyRequest(launchRequest));
+        app.on('response', (handleRequest: HandleRequest) => {
+            expect(handleRequest.jovo!.$response!.isTell('ON_REQUEST')).toBeTruthy();
+            done();
+        });
+    });
+
+    test('ON_REQUEST skip intent handling inside of a callback', async (done) => {
+
+        app.setHandler({
+            NEW_SESSION(jovo, callback) {
+                setTimeout(() => {
+                    this.tell('ON_REQUEST').skipIntentHandling();
+                    callback!();
+                }, 5);
+            },
+            LAUNCH() {
+                // skip this
+                this.tell('LAUNCH');
+            },
+        });
+        const launchRequest:JovoRequest = await t.requestBuilder.launch();
+
+        app.handle(ExpressJS.dummyRequest(launchRequest));
+        app.on('response', (handleRequest: HandleRequest) => {
+            expect(handleRequest.jovo!.$response!.isTell('ON_REQUEST')).toBeTruthy();
+            done();
+        });
+    });
+
 });
 describe('test handleOnNewSession', () => {
     test('no NEW_SESSION', async (done) => {
@@ -816,6 +860,50 @@ describe('test handleOnNewSession', () => {
         app.handle(ExpressJS.dummyRequest(launchRequest));
 
     });
+
+    test('NEW_SESSION return and skip intent handling', async (done) => {
+
+        app.setHandler({
+            NEW_SESSION() {
+                return this.tell('NEW_SESSION');
+            },
+            LAUNCH() {
+                // skip this
+                this.tell('LAUNCH');
+            },
+        });
+        const launchRequest:JovoRequest = await t.requestBuilder.launch();
+
+        app.handle(ExpressJS.dummyRequest(launchRequest));
+        app.on('response', (handleRequest: HandleRequest) => {
+            expect(handleRequest.jovo!.$response!.isTell('NEW_SESSION')).toBeTruthy();
+            done();
+        });
+    });
+
+    test('NEW_SESSION skip intent handling inside of a callback', async (done) => {
+
+        app.setHandler({
+            NEW_SESSION(jovo, callback) {
+                setTimeout(() => {
+                    this.tell('NEW_SESSION').skipIntentHandling();
+                    callback!();
+                }, 5);
+            },
+            LAUNCH() {
+                // skip this
+                this.tell('LAUNCH');
+            },
+        });
+        const launchRequest:JovoRequest = await t.requestBuilder.launch();
+
+        app.handle(ExpressJS.dummyRequest(launchRequest));
+        app.on('response', (handleRequest: HandleRequest) => {
+            expect(handleRequest.jovo!.$response!.isTell('NEW_SESSION')).toBeTruthy();
+            done();
+        });
+    });
+
 });
 
 describe('test handleOnNewUser', () => {
@@ -917,6 +1005,50 @@ describe('test handleOnNewUser', () => {
         app.handle(ExpressJS.dummyRequest(launchRequest.setUserId(randomUserId())));
 
     });
+
+    test('NEW_USER return and skip intent handling', async (done) => {
+
+        app.setHandler({
+            NEW_USER() {
+                return this.tell('NEW_USER');
+            },
+            LAUNCH() {
+                // skip this
+                this.tell('LAUNCH');
+            },
+        });
+        const launchRequest:JovoRequest = await t.requestBuilder.launch();
+
+        app.handle(ExpressJS.dummyRequest(launchRequest.setUserId(randomUserId())));
+        app.on('response', (handleRequest: HandleRequest) => {
+            expect(handleRequest.jovo!.$response!.isTell('NEW_USER')).toBeTruthy();
+            done();
+        });
+    });
+
+    test('NEW_USER skip intent handling inside of a callback', async (done) => {
+
+        app.setHandler({
+            NEW_USER(jovo, callback) {
+                setTimeout(() => {
+                    this.tell('NEW_USER').skipIntentHandling();
+                    callback!();
+                }, 5);
+            },
+            LAUNCH() {
+                // skip this
+                this.tell('LAUNCH');
+            },
+        });
+        const launchRequest:JovoRequest = await t.requestBuilder.launch();
+
+        app.handle(ExpressJS.dummyRequest(launchRequest.setUserId(randomUserId())));
+        app.on('response', (handleRequest: HandleRequest) => {
+            expect(handleRequest.jovo!.$response!.isTell('NEW_USER')).toBeTruthy();
+            done();
+        });
+    });
+
 });
 describe('test NEW_USER + NEW_SESSION + ON_REQUEST', () => {
 
@@ -1092,6 +1224,8 @@ describe('test app listener', () => {
     });
 
 });
-function randomUserId() {
+
+
+const randomUserId = () => {
     return 'user-' + Math.random().toString(36).substring(5) + '-' + Math.random().toString(36).substring(2);
-}
+};